Refactor OrganizationEdit component
